src/api/requirements.txt
fastapi[standard]
uvicorn
pydantic-settings
langchain
langchain-ollama
langchain-community